Add comprehensive tests for ingest, transform, and load modules
Introduces new test files for the weather data ETL pipeline: test_ingest.py, test_integration.py, test_load.py, and test_transform.py. These tests cover unit and integration scenarios for data ingestion, transformation, loading, error handling, data quality, and performance, ensuring robust validation of all major pipeline components.

+++ src/ingest.py
@@ -31,7 +31,7 @@ class WeatherDataPoint:
     wind_speed: float
     wind_direction: int
     cloudiness: int
-    visibility: int
+    visibility: float
     lat: float
     lon: float
